+++ packages/extension-koni-base/src/api/tokens/wasm/utils.ts
@@ -8,16 +8,73 @@ import { PSP22Contract, PSP34Contract, ShidenPSP34Contract } from '@subwallet/ex
 
 import { ApiPromise } from '@polkadot/api';
 import { Abi, ContractPromise } from '@polkadot/api-contract';
+import { WeightV2 } from '@polkadot/types/interfaces';
+import { Codec } from '@polkadot/types/types';
+import { BN } from '@polkadot/util';
 
-export const CHAINS_WITH_WEIGHT_V2 = ['shiden', 'astar', 'shibuya'];
+// @ts-ignore
+const MAX_CALL_WEIGHT = '5000000000000';
+const DEFAULT_REF_TIME = '1000000000000';
 
-export function getWasmContractGasLimit (chain: string, apiPromise: ApiPromise) {
-  return { gasLimit: CHAINS_WITH_WEIGHT_V2.includes(chain) ? getWeightV2(apiPromise) : -1 };
+export interface WasmContractResponse {
+  ok: any;
 }
 
-export function getWeightV2 (apiPromise: ApiPromise) {
-  const proofSize = 3407872;
-  const refTime = 32490000000;
+const toContractAbiMessage = (contractPromise: ContractPromise, message: string) => {
+  const value = contractPromise.abi.messages.find((m) => m.method === message);
+
+  if (!value) {
+    const messages = contractPromise?.abi.messages
+      .map((m) => m.method)
+      .join(', ');
+
+    const error = `"${message}" not found in metadata.spec.messages: [${messages}]`;
+
+    console.error(error);
+
+    return { ok: false, error };
+  }
+
+  return { ok: true, value };
+};
+
+export async function getWasmContractGasLimit (
+  api: ApiPromise,
+  callerAddress: string,
+  message: string,
+  contract: ContractPromise,
+  options = {},
+  args = []
+) {
+  try {
+    const abiMessage = toContractAbiMessage(contract, message);
+
+    if (!abiMessage.ok) {
+      return getDefaultWeightV2(api, true);
+    }
+
+    // @ts-ignore
+    const { gasLimit, storageDepositLimit, value } = options;
+
+    // @ts-ignore
+    const { gasRequired } = await api.call.contractsApi.call(
+      callerAddress,
+      contract.address,
+      value ?? new BN(0),
+      gasLimit ?? null,
+      storageDepositLimit ?? null,
+      abiMessage?.value?.toU8a(args)
+    );
+
+    return gasRequired as Codec;
+  } catch {
+    return getDefaultWeightV2(api, true);
+  }
+}
+
+export function getDefaultWeightV2 (apiPromise: ApiPromise, isFallback?: boolean): WeightV2 {
+  const proofSize = isFallback ? 3407872 : MAX_CALL_WEIGHT; // TODO: handle error better
+  const refTime = isFallback ? 32490000000 : DEFAULT_REF_TIME;
 
   return apiPromise.registry.createType('WeightV2', {
     refTime,
@@ -37,9 +94,9 @@ export async function validateWasmToken (chain: string, contractAddress: string,
       tokenContract = new ContractPromise(apiPromise, PSP22Contract, contractAddress);
 
       const [nameResp, symbolResp, decimalsResp] = await Promise.all([
-        tokenContract.query['psp22Metadata::tokenName'](contractCaller || contractAddress, { gasLimit: -1 }), // read-only operation so no gas limit
-        tokenContract.query['psp22Metadata::tokenSymbol'](contractCaller || contractAddress, { gasLimit: -1 }),
-        tokenContract.query['psp22Metadata::tokenDecimals'](contractCaller || contractAddress, { gasLimit: -1 })
+        tokenContract.query['psp22Metadata::tokenName'](contractCaller || contractAddress, { gasLimit: getDefaultWeightV2(apiPromise) }), // read-only operation so no gas limit
+        tokenContract.query['psp22Metadata::tokenSymbol'](contractCaller || contractAddress, { gasLimit: getDefaultWeightV2(apiPromise) }),
+        tokenContract.query['psp22Metadata::tokenDecimals'](contractCaller || contractAddress, { gasLimit: getDefaultWeightV2(apiPromise) })
       ]);
 
       if (!(nameResp.result.isOk && symbolResp.result.isOk && decimalsResp.result.isOk) || !nameResp.output || !decimalsResp.output || !symbolResp.output) {
@@ -52,13 +109,19 @@ export async function validateWasmToken (chain: string, contractAddress: string,
           contractError: true
         };
       } else {
-        name = symbolResp.output?.toHuman() as string;
-        decimals = parseInt(decimalsResp.output?.toHuman() as string);
-        symbol = symbolResp.output?.toHuman() as string;
+        const symbolObj = symbolResp.output?.toHuman() as Record<string, any>;
+        const decimalsObj = decimalsResp.output?.toHuman() as Record<string, any>;
+        const nameObj = nameResp.output?.toHuman() as Record<string, any>;
+
+        name = nameResp.output ? (nameObj.Ok as string || nameObj.ok as string) : '';
+        decimals = decimalsResp.output ? (new BN((decimalsObj.Ok || decimalsObj.ok) as string | number)).toNumber() : 0;
+        symbol = decimalsResp.output ? (symbolObj.Ok as string || symbolObj.ok as string) : '';
 
         if (name === '' || symbol === '') {
           contractError = true;
         }
+
+        console.log('validate PSP22', name, symbol, decimals);
       }
     } else {
       if (['astar', 'shiden', 'shibuya'].includes(chain)) {
@@ -70,7 +133,7 @@ export async function validateWasmToken (chain: string, contractAddress: string,
       }
 
       // @ts-ignore
-      const collectionIdResp = await tokenContract.query['psp34::collectionId'](contractCaller || contractAddress, getWasmContractGasLimit(chain, apiPromise)); // read-only operation so no gas limit
+      const collectionIdResp = await tokenContract.query['psp34::collectionId'](contractCaller || contractAddress, { gasLimit: getDefaultWeightV2(apiPromise) }); // read-only operation so no gas limit
 
       if (!collectionIdResp.result.isOk || !collectionIdResp.output) {
         console.error('Error response while validating WASM contract');